DES up and running in .py
Made it possible to run des in .py file. Added the boolean choice for fifo or not and the input variable for the service time distribution. Made waiting_list a global variable. Changed rho to an input variable and lambda is calculated in the function. Changed variable names to their greek counterparts in the assignment for clearity (except for lambda as this is a python function)